refactor: move files & add StringUtils
