--- app/src/main/java/kr/ksw/mybudget/presentation/core/common/IntToPriceString.kt
@@ -1,6 +1,0 @@
-package kr.ksw.mybudget.presentation.core.common
-
-import java.text.DecimalFormat
-
-fun Int.toPriceString(): String =
-    DecimalFormat("###,###,###,###").format(this)
