+++ app/src/main/java/kr/ksw/mybudget/presentation/core/common/StringUtils.kt
@@ -1,11 +1,15 @@
 package kr.ksw.mybudget.presentation.core.common
 
+import java.text.DecimalFormat
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 
 const val DATE_FORMAT_YMD = "yyyy.MM.dd"
 const val DATE_FORMAT_YMD_E = "yyyy.MM.dd (EE)"
 const val DATE_FORMAT_YMD_ADD = "yyyy/MM/dd EE요일"
+
+fun Int.toPriceString(): String =
+    DecimalFormat("###,###,###,###").format(this)
 
 fun LocalDate.toDisplayString(
     formatter: String
